requirements.txt
python-dotenv~=1.0.1
moviepy~=1.0.3
openai~=1.23.6
pytube~=15.0.0
selenium~=4.20.0
whisper~=1.1.10